refactor: remove exec-server MCP shell tool support
