--- codex-rs/exec-server/BUILD.bazel
@@ -1,11 +1,0 @@
-load("//:defs.bzl", "codex_rust_crate")
-
-codex_rust_crate(
-    name = "exec-server",
-    crate_name = "codex_exec_server",
-    integration_deps_extra = ["//codex-rs/exec-server/tests/common:common"],
-    test_tags = ["no-sandbox"],
-    extra_binaries = [
-        "//codex-rs/cli:codex",
-    ],
-)
