--- codex-rs/exec-server/src/lib.rs
@@ -1,5 +1,0 @@
-#[cfg(unix)]
-mod unix;
-
-#[cfg(unix)]
-pub use unix::*;
