--- codex-rs/exec-server/tests/all.rs
@@ -1,3 +1,0 @@
-// Single integration test binary that aggregates all test modules.
-// The submodules live in `tests/suite/`.
-mod suite;
