--- codex-rs/exec-server/tests/common/BUILD.bazel
@@ -1,7 +1,0 @@
-load("//:defs.bzl", "codex_rust_crate")
-
-codex_rust_crate(
-    name = "common",
-    crate_name = "exec_server_test_support",
-    crate_srcs = glob(["*.rs"]),
-)
