--- codex-rs/exec-server/tests/common/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "exec_server_test_support"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[lib]
-path = "lib.rs"
-
-[dependencies]
-anyhow = { workspace = true }
-codex-core = { workspace = true }
-codex-protocol = { workspace = true }
-codex-utils-cargo-bin = { workspace = true }
-rmcp = { workspace = true }
-serde_json = { workspace = true }
-tokio = { workspace = true }
